requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
bip39==0.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
mnemonic==0.20
multidict==6.0.2
pycparser==2.21
PyNaCl==1.5.0
pytz==2022.1
requests==2.27.1
urllib3==1.26.9
yarl==1.7.2
